requirements.txt
asgiref==3.8.1
async-timeout==4.0.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
crispy-bootstrap5==2024.2
cryptography==42.0.7
dj-database-url==2.2.0
Django==5.0.6
django-allauth==0.63.3
django-cors-headers==4.4.0
django-crispy-forms==2.1
django-environ==0.11.2
django-filter==24.2
django-phonenumber-field==7.3.0
djangorestframework==3.15.1
gunicorn==22.0.0
idna==3.7
packaging==24.1
phonenumbers==8.13.38
pillow==10.3.0
psycopg2-binary==2.9.9
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
redis==5.0.6
requests==2.32.3
sqlparse==0.5.0
typing_extensions==4.12.1
tzdata==2024.1
urllib3==2.2.1
whitenoise==6.7.0
